refactor

+++ LibraryManagementSystem/src/main/java/com/example/LibraryManagementSystem/dto/CategoryWithAvgPages.java
@@ -4,6 +4,6 @@ import com.example.LibraryManagementSystem.entity.Category;
 
 public record CategoryWithAvgPages(
         Category category,
-        Double avgPages
+        Integer avgPages
 ) {
 }
